fix(rbac): move audit log to separated component
